requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-pandas==0.6.7
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
numpy==2.1.1
packaging==24.2
pandas==2.2.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.32.1
whitenoise==6.8.2
